Removed unused, renamed stuff
