--- source/TODO.md
@@ -1,8 +1,0 @@
-# TODO
-- Improve documentation
-- Separate GUI into classes
-- compiler option `"strictNullChecks": true `
-- compiler option `"noImplicitAny": true,`
-- compiler option `"noImplicitReturns" : true,`
-- compiler option `"noImplicitThis": true,`
-- More unittests
