+++ source/contributing.md
@@ -1,7 +1,8 @@
-# Development Environment
-- install nodejs 
-- install npm
-- run `npm ci`
+# Getting started
+1. `git clone git@github.com:TGlas/tscript.git`
+2. install nodejs and npm
+3. run `npm ci`
 
-# Compiling
-To compile and continuously check for changes in the source run: `npm run dev`.
+- To build run `npm run build`
+- To run the unittest `npm test`
+- To watch for changes and compile run `npm run dev`
