+++ source/package.json
@@ -24,7 +24,6 @@
     "ts-loader": "^8.0.14",
     "ts-node": "^9.1.1",
     "typescript": "^4.1.3",
-    "uglifyjs-webpack-plugin": "^2.2.0",
     "webpack": "^5.15.0",
     "webpack-cli": "^4.3.1"
   },
