+++ source/src/parse/parse_call.ts
@@ -1,6 +1,6 @@
 import { ErrorHelper } from "../errors/ErrorHelper";
 import { Lexer } from "./lexer";
-import { get_type, copyconstant } from "../interpreter/runner_helper";
+import { get_type, copyconstant } from "../interpreter/interpreter_helper";
 import { TScript } from "../tscript";
 import { simfalse } from "../helpers/sims";
 import { Typeid } from "../helpers/typeIds";
