+++ source/src/parse/parse_name.ts
@@ -8,7 +8,7 @@
 import { ErrorHelper } from "../errors/ErrorHelper";
 import { Lexer } from "./lexer";
 import { Options } from "../helpers/options";
-import { get_function, get_type } from "../interpreter/runner_helper";
+import { get_function, get_type } from "../interpreter/interpreter_helper";
 import { TScript } from "../tscript";
 import { resolve_name } from "./parse_fn";
