+++ source/src/version.ts
@@ -1,13 +1,12 @@
 export const Version = {
-    type: "beta",
-    major: 0,
-    minor: 5,
-    patch: 43,
-    day: 7,
-    month: 1,
-    year: 2021,
-    full: function()
-            {
+        type: "beta",
+        major: 0,
+        minor: 5,
+        patch: 43,
+        day: 7,
+        month: 1,
+        year: 2021,
+        full: function () {
                 let s = "TScript version " + this.major
                         + "." + this.minor
                         + "." + this.patch;
@@ -17,5 +16,5 @@ export const Version = {
                         + "." + this.year
                         + " - (C) Tobias Glasmachers 2018" + "-" + this.year;
                 return s;
-            },
-    }
+        },
+}
